Bobbranch

+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+
 use \App\User;
 use \App\Account;
 use \App\Input;
@@ -47,12 +49,15 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $user = User::findOrFail($id);
+        if (Auth::user()->id==$id) {
+            $user = User::findOrFail($id);
 
-        return view('account.edit', [
-            'user' => $user
-        ]);
-
+            return view('account.edit', [
+                'user' => $user
+            ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -64,22 +69,25 @@ class UserController extends Controller
      */
     public function update(user $user, Request $request, $id)
     {
-        $request->photo->store('public/photos');
+        if (Auth::user()->id==$id) {
+            $request->photo->storeAs('public/profilepics', 'profilepic' . $id .'.jpg');
 
-        $user = User::findOrFail($id);
-        // $user->update(request(['name', 'email', 'phone', 'photo', 'skills']));
+            $user = User::findOrFail($id);
+            // $user->update(request(['name', 'email', 'phone', 'photo', 'skills']));
 
-        $user->name = request('name');
-        $user->email = request('email');
-        $user->phone = request('phone');
-        $user->skills = request('skills');
-        $user->about_me = request('about_me');
+            $user->name = request('name');
+            $user->email = request('email');
+            $user->phone = request('phone');
+            $user->skills = request('skills');
+            $user->about_me = request('about_me');
+            
+            $user->update();
+
+            return redirect('/account/' . $id);
+        } else {
+            return redirect('/');
+        }
         
-        
-
-        $user->update();
-
-        return redirect('/account/' . $id);
     }
 
     /**
@@ -90,20 +98,13 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if (Auth::user()->id==$id) {
         $user = User::findOrFail($id);
 
         $user->delete();
+        }
 
         return redirect('/');
     }
 
-    public function uploadPhoto (Request $request, $id) 
-    {
-                // dd($request->photo);
-
-            $request->photo->store('photos');
-            //  dd($request->photo);
-            // dd($request->hasFile('photo'));
-            return redirect('/account');
-    }
 }
